requirements.txt
Flask>=2.3.2
Flask-RESTful>=0.3.10
pytest>=7.4.0
mysqlclient>=2.2.0
SQLAlchemy>=2.0.19
python-dotenv>=1.0.0